BotUser id изменен на long

+++ src/main/java/com/github/l524l/telegramnekobot/repositories/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<BotUser, Integer> {
+public interface UserRepository extends JpaRepository<BotUser, Long> {
     Optional<BotUser> findBotUserByUsername(String name);
 }
